+++ src/cover_crypt_core.rs
@@ -318,7 +318,7 @@ impl Encapsulation {
 
 /// CovCrypt secret key is a vector of bytes of the same length as secret key
 /// of the underlying KEM.
-#[derive(Clone, PartialEq, Debug)]
+#[derive(Clone, PartialEq, Eq, Debug)]
 pub struct SecretKey(Vec<u8>);
 
 impl std::ops::Deref for SecretKey {