+++ src/interfaces/pyo3/generate_cc_keys.rs
@@ -3,15 +3,8 @@ use pyo3::{exceptions::PyTypeError, prelude::*};
 
 use crate::{
     api::{CoverCrypt, PrivateKey},
-    error::Error,
     policies::{AccessPolicy, Policy},
 };
-
-impl From<Error> for pyo3::PyErr {
-    fn from(e: Error) -> Self {
-        pyo3::exceptions::PyTypeError::new_err(format!("{e}"))
-    }
-}
 
 #[pyfunction]
 pub fn generate_master_keys(policy_bytes: Vec<u8>) -> PyResult<(Vec<u8>, Vec<u8>)> {
@@ -36,8 +29,8 @@ pub fn generate_user_private_key(
     policy_bytes: Vec<u8>,
 ) -> PyResult<Vec<u8>> {
     let master_private_key: PrivateKey<X25519Crypto> =
-        PrivateKey::try_from_bytes(master_private_key_bytes.to_vec().as_slice())?;
-    let policy = serde_json::from_slice(policy_bytes.to_vec().as_slice())
+        PrivateKey::try_from_bytes(&master_private_key_bytes)?;
+    let policy = serde_json::from_slice(&policy_bytes)
         .map_err(|e| PyTypeError::new_err(format!("Policy deserialization failed: {e}")))?;
     let access_policy = AccessPolicy::from_boolean_expression(&access_policy_str)?;